requirements.txt
ccxt>=4.0.0
ccxt[pro]
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b
tensorflow>=2.15.0
optuna>=3.5.0
joblib>=1.3.0
python-telegram-bot>=20.0
streamlit>=1.30.0
plotly>=5.18.0
aiosqlite>=0.19.0
aiohttp>=3.9.0
scikit-learn>=1.3.0
scipy>=1.11.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
ta>=0.11.0
